http: support wrapping handler function
