+++ wrappers/net/http/handler_wrapper_test.go
@@ -1,48 +1,39 @@
-package epsagongin
+package epsagonhttp
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
 	"io/ioutil"
+	"net/http"
 	"net/http/httptest"
 	"testing"
 
 	"github.com/epsagon/epsagon-go/epsagon"
 	"github.com/epsagon/epsagon-go/protocol"
 	"github.com/epsagon/epsagon-go/tracer"
-	"github.com/gin-gonic/gin"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
 
-func TestGinWrapper(t *testing.T) {
+func TestHandlerWrapper(t *testing.T) {
 	RegisterFailHandler(Fail)
 	RunSpecs(t, "Gin Wrapper")
 }
 
-type MockEngine struct {
-	*gin.Engine
-	TestHandler func(handler gin.HandlerFunc)
-}
-
-func (me *MockEngine) Handle(httpMethod, relativePath string, handlers ...gin.HandlerFunc) gin.IRoutes {
-	me.TestHandler(handlers[0])
-	return nil
-}
-
 var _ = Describe("gin_wrapper", func() {
-	Describe("GinRouterWrapper", func() {
+	Describe("WrapHandleFunc", func() {
 		var (
 			events         []*protocol.Event
 			exceptions     []*protocol.Exception
-			wrapper        *GinRouterWrapper
-			mockedEngine   *MockEngine
+			request        *http.Request
+			responseWriter *httptest.ResponseRecorder
 			called         bool
-			testGinContext *gin.Context
+			config         *epsagon.Config
 		)
 		BeforeEach(func() {
 			called = false
-			config := &epsagon.Config{Config: tracer.Config{
+			config = &epsagon.Config{Config: tracer.Config{
 				Disable:  true,
 				TestMode: true,
 			}}
@@ -54,51 +45,45 @@ var _ = Describe("gin_wrapper", func() {
 				Labels:     make(map[string]interface{}),
 				Config:     &config.Config,
 			}
-			mockedEngine = &MockEngine{
-				Engine: gin.New(),
-				TestHandler: func(handler gin.HandlerFunc) {
-					handler(testGinContext)
-				},
-			}
-			wrapper = &GinRouterWrapper{
-				IRouter:  mockedEngine,
-				Hostname: "test",
-				Config:   config,
-			}
 			body := []byte("hello")
-			testGinContext, _ = gin.CreateTestContext(httptest.NewRecorder())
-			testGinContext.Request = httptest.NewRequest("POST", "https://www.help.com", ioutil.NopCloser(bytes.NewReader(body)))
-			Expect(testGinContext.Request).NotTo(Equal(nil))
+			request = httptest.NewRequest("POST", "https://www.help.com", ioutil.NopCloser(bytes.NewReader(body)))
+			responseWriter = httptest.NewRecorder()
 		})
 		Context("Happy Flows", func() {
 			It("calls the wrapped function", func() {
-				mockedEngine.TestHandler = func(handler gin.HandlerFunc) {
-					handler(testGinContext)
-					Expect(called).To(Equal(true))
-				}
-				wrapper.GET("/test", func(c *gin.Context) { called = true })
+				wrapper := WrapHandleFunc(
+					config,
+					func(rw http.ResponseWriter, req *http.Request) {
+						called = true
+					},
+				)
+				wrapper(responseWriter, request)
+				Expect(called).To(Equal(true))
 			})
-			It("passes the tracer through gin context", func() {
-				mockedEngine.TestHandler = func(handler gin.HandlerFunc) {
-					handler(testGinContext)
-					Expect(called).To(Equal(true))
-				}
-				wrapper.GET("/test", func(c *gin.Context) {
-					tracer := c.Keys[TracerKey].(tracer.Tracer)
-					tracer.AddLabel("test", "ok")
-					called = true
-				})
+			It("passes the tracer through request context", func() {
+				wrapper := WrapHandleFunc(
+					config,
+					func(rw http.ResponseWriter, req *http.Request) {
+						called = true
+						tracer := epsagon.ExtractTracer([]context.Context{req.Context()})
+						tracer.AddLabel("test", "ok")
+					},
+				)
+				wrapper(responseWriter, request)
+				Expect(called).To(Equal(true))
 				Expect(
 					tracer.GlobalTracer.(*tracer.MockedEpsagonTracer).Labels["test"],
 				).To(Equal("ok"))
 			})
 			It("Creates a runner and trigger events for handler invocation", func() {
-				mockedEngine.TestHandler = func(handler gin.HandlerFunc) {
-					handler(testGinContext)
-				}
-				wrapper.GET("/test", func(c *gin.Context) {
-					called = true
-				})
+				wrapper := WrapHandleFunc(
+					config,
+					func(rw http.ResponseWriter, req *http.Request) {
+						called = true
+					},
+					"test-handler",
+				)
+				wrapper(responseWriter, request)
 				Expect(len(events)).To(Equal(2))
 				var runnerEvent *protocol.Event
 				for _, event := range events {
@@ -107,25 +92,38 @@ var _ = Describe("gin_wrapper", func() {
 					}
 				}
 				Expect(runnerEvent).NotTo(Equal(nil))
-				Expect(runnerEvent.Resource.Type).To(Equal("gin"))
-				Expect(runnerEvent.Resource.Name).To(Equal("/test"))
+				Expect(runnerEvent.Resource.Type).To(Equal("go-function"))
+				Expect(runnerEvent.Resource.Name).To(Equal("test-handler"))
 			})
 			It("Adds correct trigger event", func() {
 				body := []byte("hello world")
-				testGinContext.Request = httptest.NewRequest(
+				request = httptest.NewRequest(
 					"POST",
 					"https://www.help.com/test?hello=world&good=bye",
 					ioutil.NopCloser(bytes.NewReader(body)))
-				wrapper.Hostname = ""
-				wrapper.GET("/test", func(c *gin.Context) {
-					internalHandlerBody, err := ioutil.ReadAll(c.Request.Body)
-					if err != nil {
-						Expect(true).To(Equal(false))
-					}
-					Expect(internalHandlerBody).To(Equal(body))
-					called = true
-					c.JSON(200, gin.H{"hello": "world"})
-				})
+				wrapper := WrapHandleFunc(
+					config,
+					func(rw http.ResponseWriter, req *http.Request) {
+						called = true
+						internalHandlerBody, err := ioutil.ReadAll(req.Body)
+						if err != nil {
+							Expect(true).To(Equal(false))
+						}
+						Expect(internalHandlerBody).To(Equal(body))
+						resp, err := json.Marshal(map[string]string{"hello": "world"})
+						if err != nil {
+							Expect(true).To(Equal(false))
+						}
+						rw.Header().Add("Content-Type", "application/json; charset=utf-8")
+						_, err = rw.Write(resp)
+						if err != nil {
+							Expect(true).To(Equal(false))
+						}
+
+					},
+					"test-handler",
+				)
+				wrapper(responseWriter, request)
 				Expect(len(events)).To(Equal(2))
 				var triggerEvent *protocol.Event
 				for _, event := range events {
@@ -157,11 +155,18 @@ var _ = Describe("gin_wrapper", func() {
 			It("Adds Exception if handler explodes", func() {
 				errorMessage := "boom"
 				Expect(func() {
-					wrapper.GET("/test", func(c *gin.Context) {
-						panic(errorMessage)
-					})
+					wrapper := WrapHandleFunc(
+						config,
+						func(rw http.ResponseWriter, req *http.Request) {
+							called = true
+							panic(errorMessage)
+						},
+						"test-handler",
+					)
+					wrapper(responseWriter, request)
 				}).To(
 					PanicWith(epsagon.MatchUserError(errorMessage)))
+				Expect(called).To(Equal(true))
 				Expect(len(events)).To(Equal(2))
 				var runnerEvent, triggerEvent *protocol.Event
 				for _, event := range events {
@@ -172,9 +177,6 @@ var _ = Describe("gin_wrapper", func() {
 						triggerEvent = event
 					}
 				}
-				Expect(
-					tracer.GlobalTracer.(*tracer.MockedEpsagonTracer).Stopped(),
-				).To(Equal(true))
 				Expect(runnerEvent.Exception).NotTo(Equal(nil))
 				Expect(triggerEvent.Exception).NotTo(Equal(nil))
 				Expect(triggerEvent.Resource.Metadata["status_code"]).To(
